Fix type inference issue related to nullabe arrays
If a JSON element array is missing from one instance of the owning element, the array was considered to have nullable elements. An element array should only have nullable elements if `null` is found in the array. The logic was updated to carry over the known nullability evidence, instead of letting the absence of an array influence the detemination of whether it can contain `null` elements.